+ve/-ve positions

+++ iotests/positions.c
@@ -102,7 +102,7 @@ void createLinearPositions(positionType *positions,
 			   const size_t bdSizeBytes,     // max block device size in bytes
 			   const size_t startBlock, // start block * bdSize
 			   const size_t finBlock,   // finis block * bdSize
-			   const size_t jumpBlock,
+			   const int    jumpBlock,
 			   const double rwRatio) {
 
   if (bdSizeBytes < blockSize) {
@@ -140,7 +140,7 @@ void createParallelPositions(positionType *positions,
 			     const size_t numPositions,
 			     const size_t blockSize,
 			     const size_t bdSizeBytes,
-			     const size_t jumpBlock,
+			     const int    jumpBlock,
 			     const double rwRatio,
 			     const size_t parallelRegions) {
 
@@ -220,7 +220,7 @@ void setupPositions(positionType *positions,
 			  const double readorwrite,
 			  const size_t bs,
 			  const size_t singlePosition,
-		          const size_t jumpStep,
+		          const int    jumpStep,
 		          const size_t startAtZero
 		    ) {
   if (bdSizeBytes < bs) {
