+++ iotests/positions.h
@@ -19,14 +19,14 @@ void createLinearPositions(positionType *positions,
 			   const size_t bdSizeBytes,     // max block device size in bytes
 			   const size_t startBlock, // start block * bdSize
 			   const size_t finBlock,   // finis block * bdSize
-			   const size_t jumpBlock,
+			   const int    jumpBlock,
 			   const double rwRatio);
 
 void createParallelPositions(positionType *positions,
 			     const size_t numPositions,
 			     const size_t blockSize,
 			     const size_t bdSizeBytes,
-			     const size_t jumpBlock,
+			     const int    jumpBlock,
 			     const double rwRatio,
 			     const size_t parallelRegions);
 
@@ -37,7 +37,7 @@ void setupPositions(positionType *positions,
 		    const double readorwrite,
 		    const size_t bs,
 		    const size_t singlePosition,
-		    const size_t jumpStep,
+		    const int    jumpStep,
 		    const size_t startAtZero);
